refactor: 사용자 컴포넌트 분리
