+++ moigoFront/src/screens/user/Home/EventCard.tsx
