+++ moigoFront/src/screens/user/Meeting/MeetingModal.tsx
